% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyToEnglish.R
\name{GCD}
\alias{GCD}
\title{Greater Common Denominator}
\usage{
GCD(...)
}
\arguments{
\item{...}{Any number of `numeric` vectors or nested `list`s containing such}
}
\value{
A `numeric` that is the greatest common denominator of the input values
}
\description{
Calculates greatest common denominator of a list of numbers
}
