% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyToEnglish.R
\name{generate_random_sentences}
\alias{generate_random_sentences}
\title{Generate Random Sentences}
\usage{
generate_random_sentences(n, punctuate = TRUE, fast = FALSE)
}
\arguments{
\item{n}{`numeric` number of sentences to generate}

\item{punctuate}{`logical` value of whether to add spaces, capitalize first letter, and append period}

\item{fast}{`logical`}
}
\value{
`character` vector of randomly generated sentences
}
\description{
Randomly generate sentences with a specific structure
}
