% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyToEnglish.R
\name{hash_to_sentence}
\alias{hash_to_sentence}
\title{Hash to Sentence}
\usage{
hash_to_sentence(x, ...)
}
\arguments{
\item{x}{- Input data, which will be converted to `character` if not already `character`}

\item{...}{- Other parameters to pass to `keyToEnglish()`,
besides `word_list`, `hash_subsection_size`, and `hash_function`}
}
\value{
`character` vector of hashed field resembling phrases
}
\description{
Hashes data to a sentence that contains 54 bits of entropy
}
