% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyToEnglish.R
\name{LCM}
\alias{LCM}
\title{Least Common Multiple}
\usage{
LCM(...)
}
\arguments{
\item{...}{Any number of `numeric` vectors or nested `list`s containing such}
}
\value{
A `numeric` that is the least common multiple of the input values
}
\description{
Calculates least common multiple of a list of numbers
}
