% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile_misspellings}
\alias{reconcile_misspellings}
\title{Reconcile Misspellings}
\usage{
reconcile_misspellings(x, from = "0.2.0", to = "0.2.1")
}
\arguments{
\item{x}{`character` vector of phrases from `keyToEnglish()` output or related function}

\item{from}{`character` version number of how the hash was originally generated}

\item{to}{`character` version that hash should be converted to}
}
\value{
`character` vector of fixed phrase hash compatible with output for version `to`
}
\description{
Fixes incorrect spellings from previous versions of this package
}
