% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyToEnglish.R
\name{uniqueness_probability}
\alias{uniqueness_probability}
\title{Uniqueness Probability}
\usage{
uniqueness_probability(N, r)
}
\arguments{
\item{N}{`numeric` size of set. Becomes unstable for values greater than 10^16.}

\item{r}{`numeric` number of elements selected with replacement}
}
\value{
`numeric` probability that all `r` elements are unique
}
\description{
Calculates probability that all `r` elements of a set of size `N` are unique
}
