% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_builder.R
\name{wiki_clean}
\alias{wiki_clean}
\title{Clean JSON text from Wikipedia}
\usage{
wiki_clean(x)
}
\arguments{
\item{x}{`character` JSON text}
}
\value{
`character` JSON text
}
\description{
Clean JSON text from Wikipedia
}
