% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_adjectives_nonorigin}
\alias{wl_adjectives_nonorigin}
\title{Non-Origin Adjectives Wordlist}
\format{
A `character` vector
}
\usage{
data(wl_adjectives_nonorigin)
}
\description{
Word list of 256 adjectives that do not describe origin, so they can
usually be used prior to visual/origin adjectives without breaking
any grammar rules
}
\keyword{datasets}
