% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_adjectives_visual}
\alias{wl_adjectives_visual}
\title{Visual Adjectives Wordlist}
\format{
A `character` vector
}
\usage{
data(wl_adjectives_visual)
}
\description{
Word list of 256 adjectives that visually describe an object.
}
\keyword{datasets}
