% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_animal}
\alias{wl_animal}
\title{Animal word list}
\format{
An object of class `character`
}
\usage{
data(wl_animal)
}
\description{
Word list generated by processing several animal-related pages
on Wikipedia
}
\examples{
data(wl_animal)
keyToEnglish(1:5, word_list=wl_animal)
}
\keyword{datasets}
