% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_common}
\alias{wl_common}
\title{Common word list}
\format{
An object of class `character`
}
\usage{
data(wl_common)
}
\description{
Public domain word list of common words
}
\examples{
data(wl_common)
keyToEnglish(1:5, word_list=wl_common)
}
\references{
Public Domain Word Lists. Michael Wehar
\href{https://github.com/MichaelWehar/Public-Domain-Word-Lists}{https://github.com/MichaelWehar/Public-Domain-Word-Lists}
}
\keyword{datasets}
