% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_freq5663}
\alias{wl_freq5663}
\title{Freq 5663 word list}
\format{
An object of class `character`
}
\usage{
data(wl_freq5663)
}
\description{
Public domain word list of common words, slightly truncated
from original version
}
\examples{
data(wl_common)
keyToEnglish(1:5, word_list=wl_freq5663)
}
\references{
Public Domain Word Lists. Michael Wehar
\href{https://github.com/MichaelWehar/Public-Domain-Word-Lists}{https://github.com/MichaelWehar/Public-Domain-Word-Lists}
}
\keyword{datasets}
