% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_literature}
\alias{wl_literature}
\title{Literature word list}
\format{
An object of class `character`
}
\usage{
data(wl_literature)
}
\description{
Word list generated by processing several works of literature on
Project Gutenberg
}
\examples{
data(wl_literature)
keyToEnglish(1:5, word_list=wl_literature)
}
\references{
Project Gutenberg.
\href{https://gutenberg.org}{Project Gutenberg}
}
\keyword{datasets}
