% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_nouns_concrete}
\alias{wl_nouns_concrete}
\title{Concrete Nouns Wordlist}
\format{
A `character` vector
}
\usage{
data(wl_nouns_concrete)
}
\description{
Word list of 2048 singular, concrete nouns, largely excluding materials and liquids that
cannot be referred to in the singular form
}
\keyword{datasets}
