% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_nouns_concrete_plural}
\alias{wl_nouns_concrete_plural}
\title{Plural Concrete Nouns Wordlist}
\format{
A `character` vector
}
\usage{
data(wl_nouns_concrete_plural)
}
\description{
Word list of 2048 concrete nouns in plural form, largely excluding materials and liquids that
cannot be referred to in the singular form.
}
\keyword{datasets}
