% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_science}
\alias{wl_science}
\title{Science word list}
\format{
An object of class `character`
}
\usage{
data(wl_science)
}
\description{
Word list generated by processing several science-related pages
on Wikipedia
}
\examples{
data(wl_science)
keyToEnglish(1:5, word_list=wl_science)
}
\keyword{datasets}
