% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_verbs_transitive_gerund}
\alias{wl_verbs_transitive_gerund}
\title{Transitive Verbs in Gerund Form}
\format{
A `character` vector
}
\usage{
data(wl_verbs_transitive_gerund)
}
\description{
Word list of 256 transitive verbs in gerund form (i.e., "ing" at end)
}
\keyword{datasets}
