% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_verbs_transitive_infinitive}
\alias{wl_verbs_transitive_infinitive}
\title{Transitive Verbs in Infinitive Form}
\format{
A `character` vector
}
\usage{
data(wl_verbs_transitive_infinitive)
}
\description{
Word list of 256 transitive verbs in infinitive form (minus the "to")
}
\keyword{datasets}
