% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wl_verbs_transitive_present}
\alias{wl_verbs_transitive_present}
\title{Transitive Verbs in Present Form}
\format{
A `character` vector
}
\usage{
data(wl_verbs_transitive_present)
}
\description{
Word list of 256 transitive verbs in present tense
}
\keyword{datasets}
