% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wml_animals}
\alias{wml_animals}
\title{Animal Phrase Structure Word Multilist}
\format{
A `list` of `character` vectors

A `list` of `character` vectors
}
\usage{
data(wml_animals)

data(wml_animals)
}
\description{
Word lists of sizes, colors, animals, and attributes to construct memorable phrases

List of word lists that combine cute words with physics-related words
}
\examples{

keyToEnglish(1:5, word_list=wml_animals)
}
\keyword{datasets}
