% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wml_cutephysics}
\alias{wml_cutephysics}
\title{Cute Physics Multilist}
\format{
A `list` of `character` vectors
}
\usage{
data(wml_cutephysics)
}
\description{
List of word lists that combine cute words with physics-related words
}
\keyword{datasets}
