% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wml_long_sentence}
\alias{wml_long_sentence}
\title{Long Sentence Multilist}
\format{
A `list` of `character` vectors
}
\usage{
data(wml_long_sentence)
}
\description{
List of word lists that can be used to make a 54-byte, often humorous, sentence
}
\keyword{datasets}
